+++ gatekeeper/trafficlights.h
@@ -21,13 +21,13 @@ public:
 	bool 		get() { return m_on; }
 	operator	bool() { return m_on; }
 
-	void set( bool on, unsigned long cyclelen, uint8_t cyclecount, bool endoff, unsigned long currmillis = 0 );
+	void set( bool on, uint16_t	cyclelen, uint8_t cyclecount, bool endoff, unsigned long currmillis = 0 );
 
 private:
 	uint8_t			m_iopin;
 	bool			m_highon;
 	bool			m_on;
-	unsigned long	m_cyclelen;
+	uint16_t		m_cyclelen;
 	uint8_t			m_cyclecount;
 	bool			m_endoff;
 	unsigned long	m_lastmilli;
@@ -47,13 +47,11 @@ public:
 			trafficlight( const uint8_t pins[], bool highon );
 	bool 	init( const uint8_t pins[], bool highon );
 	void	loop( unsigned long currmillis );
-	void	set( COLORS color, bool on, unsigned long cyclelen, unsigned long currmillis = 0 );
-	void 	set( bool r, unsigned long rc, bool y, unsigned long yc, bool g, unsigned long gc, unsigned long currmillis = 0 );
+	void	set( COLORS color, bool on, uint16_t cyclelen, unsigned long currmillis = 0 );
+	void 	set( bool r, uint16_t rc, bool y, uint16_t yc, bool g, uint16_t gc, unsigned long currmillis = 0 );
 
 protected:
 	outputpin			m_lights[3];
-
-
 };
 
 //////////////////////////////////////////////////////////////////////////////
@@ -63,8 +61,8 @@ public:
 	enum STATUS { OFF=0, CODEWAIT, CONFLICT, ACCEPTED, WARNED, DENIED, PASS, NUMSTATES };
 
 	trafficlights() {}
-	trafficlights( const uint8_t innerpins[], const uint8_t outerpins[], unsigned long cyclelen );
-	bool 		init( const uint8_t innerpins[], const uint8_t outerpins[], unsigned long cyclelen );
+	trafficlights( const uint8_t innerpins[], const uint8_t outerpins[], uint16_t cyclelen );
+	bool 		init( const uint8_t innerpins[], const uint8_t outerpins[], uint16_t cyclelen );
 	void		loop( unsigned long currmillis = 0 );
 	uint16_t	set( STATUS status, bool inner );
 	void		set( uint16_t state, bool inner );
@@ -75,6 +73,6 @@ protected:
 	trafficlight			m_inner, m_outer;
 	STATUS					m_status;
 	static const uint16_t	m_compstates[NUMSTATES];
-	unsigned long			m_cyclelen;
+	uint16_t				m_cyclelen;
 };
 #endif /* TRAFFICLIGHTS_H_ */
